+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Fix Inconsistent meaning of --output=raw in get message and consume command (see fgrosse/kafkactl#2)
 - Make `github.com/fgrosse/kafkactl/pkg` internal
 - Integration with the [Confluent Schema Registry]
 - `kafkactl consume`: Support decoding Avro messages